+++ driven/storage/adapter.go
@@ -433,6 +433,18 @@ func (sa *Adapter) DeleteLoginSession(id string) error {
 	return nil
 }
 
+//DeleteExpiredSessions deletes expired sessions
+func (sa *Adapter) DeleteExpiredSessions(now *time.Time) error {
+	filter := bson.M{"expires": bson.M{"$lte": now}}
+
+	_, err := sa.db.loginsSessions.DeleteMany(filter, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeLoginSession, &logutils.FieldArgs{"expires": now}, err)
+	}
+
+	return nil
+}
+
 //FindAccount finds an account for app, org, auth type and account auth type identifier
 func (sa *Adapter) FindAccount(appID string, orgID string, authTypeID string, accountAuthTypeIdentifier string) (*model.Account, error) {
 	filter := bson.D{primitive.E{Key: "app_id", Value: appID},
@@ -922,99 +934,6 @@ func (sa *Adapter) UpdateCredential(creds *model.Credential) error {
 
 // 	return nil
 // }
-
-//FindRefreshToken finds a refresh token
-func (sa *Adapter) FindRefreshToken(token string) (*model.AuthRefresh, error) {
-	conditions := []bson.M{{"current_token": token}, {"previous_token": token}}
-	filter := bson.M{"$or": conditions}
-
-	var refresh model.AuthRefresh
-	err := sa.db.loginsSessions.FindOne(filter, &refresh, nil)
-	if err != nil {
-		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
-	}
-
-	return &refresh, nil
-}
-
-//LoadRefreshTokens loads refresh tokens by an ID triple
-func (sa *Adapter) LoadRefreshTokens(orgID string, appID string, credsID string) ([]model.AuthRefresh, error) {
-	filter := bson.D{primitive.E{Key: "org_id", Value: orgID}, primitive.E{Key: "app_id", Value: appID}, primitive.E{Key: "creds_id", Value: credsID}}
-	opts := options.Find().SetSort(bson.D{primitive.E{Key: "exp", Value: 1}})
-
-	var refresh []model.AuthRefresh
-	err := sa.db.loginsSessions.Find(filter, &refresh, opts)
-	if err != nil {
-		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
-	}
-
-	return refresh, nil
-}
-
-//InsertRefreshToken inserts a refresh token
-func (sa *Adapter) InsertRefreshToken(refresh *model.AuthRefresh) error {
-	if refresh == nil {
-		return errors.ErrorData(logutils.StatusInvalid, model.TypeAuthRefresh, nil)
-	}
-
-	_, err := sa.db.loginsSessions.InsertOne(refresh)
-	if err != nil {
-		return errors.WrapErrorAction(logutils.ActionInsert, model.TypeAuthRefresh, nil, err)
-	}
-
-	return nil
-}
-
-//UpdateRefreshToken updates a refresh token
-func (sa *Adapter) UpdateRefreshToken(token string, refresh *model.AuthRefresh) error {
-	filter := bson.D{primitive.E{Key: "current_token", Value: token}}
-	update := bson.D{
-		primitive.E{Key: "$set", Value: bson.D{
-			primitive.E{Key: "previous_token", Value: refresh.PreviousToken},
-			primitive.E{Key: "current_token", Value: refresh.CurrentToken},
-			primitive.E{Key: "exp", Value: refresh.Expires},
-			primitive.E{Key: "params", Value: refresh.Params},
-			primitive.E{Key: "date_updated", Value: refresh.DateUpdated},
-		}},
-	}
-
-	res, err := sa.db.loginsSessions.UpdateOne(filter, update, nil)
-	if err != nil {
-		return errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
-	}
-	if res.ModifiedCount != 1 {
-		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAuthRefresh, &logutils.FieldArgs{"unexpected modified count": res.ModifiedCount})
-	}
-
-	return nil
-}
-
-//DeleteRefreshToken updates a refresh token
-func (sa *Adapter) DeleteRefreshToken(token string) error {
-	filter := bson.D{primitive.E{Key: "current_token", Value: token}}
-
-	res, err := sa.db.loginsSessions.DeleteOne(filter, nil)
-	if err != nil {
-		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeAuthRefresh, nil, err)
-	}
-	if res.DeletedCount != 1 {
-		return errors.ErrorAction(logutils.ActionDelete, model.TypeAuthRefresh, logutils.StringArgs("unexpected deleted count"))
-	}
-
-	return nil
-}
-
-//DeleteExpiredRefreshTokens deletes expired refresh tokens
-func (sa *Adapter) DeleteExpiredRefreshTokens(now *time.Time) error {
-	filter := bson.M{"exp": bson.M{"$lte": now}}
-
-	_, err := sa.db.loginsSessions.DeleteMany(filter, nil)
-	if err != nil {
-		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeAuthRefresh, &logutils.FieldArgs{"exp": now}, err)
-	}
-
-	return nil
-}
 
 //FindApplicationPermissions finds a set of application permissions
 func (sa *Adapter) FindApplicationPermissions(ids []string, appID string) ([]model.ApplicationPermission, error) {