+++ driven/storage/database.go
@@ -279,8 +279,8 @@ func (m *database) applyLoginsSessionsChecks(refreshTokens *collectionWrapper) e
 func (m *database) applyAPIKeysChecks(apiKeys *collectionWrapper) error {
 	m.logger.Info("apply api keys checks.....")
 
-	// Add org_id, app_id compound index - unique
-	err := apiKeys.AddIndex(bson.D{primitive.E{Key: "org_id", Value: 1}, primitive.E{Key: "app_id", Value: 1}}, true)
+	// Add app_id index
+	err := apiKeys.AddIndex(bson.D{primitive.E{Key: "app_id", Value: 1}}, false)
 	if err != nil {
 		return err
 	}