+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement logins sessions [#172](https://github.com/rokwire/core-building-block/issues/172) 
 
 ## [1.6.0] - 2021-10-19
 ### Added
-- Implement logins sessions [#172](https://github.com/rokwire/core-building-block/issues/172) 
+- Implement logins sessions - almost completed [#172](https://github.com/rokwire/core-building-block/issues/172) 
 
 ## [1.5.0] - 2021-10-15
 ### Fixed